# Optimize decoding of struct field